Refactor create command
